refactoring polygon and math utilities
